+++ src/interface.ts
@@ -4,12 +4,14 @@ export interface IFeed {
 }
 
 export interface IEntry {
+    identifier?: string;
     title: string;
     link?: string;
     date?: Date;
 }
 
 export interface IFeedConfig {
+    identifier?: string;
 	title: string;
 	url: string;
 	id?: number;